feat: добавлены классы Friendship, Like, MpaRaiting, Genre данные хранятся в БД Н2

+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -14,5 +14,5 @@ public interface FilmStorage {
 
     List<Film> getFilmList();
 
-    Film getFilmById(int id);
+    Film getFilmById(long id);
 }
